// frontend/src/app/components/pages-sidebar/elenco-rumore/elenco-rumore-dto.ts
export class ElencoRumoreDto {
    numcodcar?: any;
    numcodcarold?: any;
    idaut?: any;
    denominazione?: string;
    gestore?: string;
    attivita?: string;
    dataril?: any;
    numautatt?: string;
    tipaut?: string;
    dataproc?: any;
    protcoll?: string;
    provincia?: string;
    comune?: any;
    coordinatelat?: string;
    coordinatelong?: string;
    indirizzo?: string;
    nomesito?: string;
    autesp?: string;
    sorgente?: string;
    tipologiaoss?: string;
    schedaobb?: string;
    note?: string;
    linkcondivisia?: string;

}